feat: implementa endpoint de delecao de transacoes em memoria

+++ src/main/java/dev/java10x/itauJava10x/TransacaoRequest.java
@@ -12,7 +12,7 @@ import java.time.OffsetDateTime;
 @NoArgsConstructor
 public class TransacaoRequest {
 
-    private BigDecimal valor;
+    private Double valor;
     private OffsetDateTime dataHora;
 
 }
